100% dataretriever test coverage
